Stabilize admin menu functionality

+++ admin_menu/application/modules/kelola_news_feed/models/Kelola_newsfeed_m.php
@@ -6,19 +6,31 @@
         }
         public function get_data_news_feed()
         {
+            if (!$this->db->table_exists('feeds')) {
+                return $this->db->query("SELECT NULL AS feedId, NULL AS subject, NULL AS gambar, NULL AS status, NULL AS create_at, NULL AS create_user WHERE 1=0");
+            }
             return $this->db->get('feeds');
         }
         public function get_news_feed($id)
         {
+            if (!$this->db->table_exists('feeds')) {
+                return array();
+            }
             $query = $this->db->get_where('feeds', ['feedId' => $id]);
             return $query->row_array();
         }
         public function insert_news($data)
         {
+            if (!$this->db->table_exists('feeds')) {
+                return false;
+            }
             $this->db->insert('feeds', $data);
         }
         public function update_news($feedId,$gambar,$subject,$status)
         {
+            if (!$this->db->table_exists('feeds')) {
+                return false;
+            }
             $data = [
                 'subject' => $subject,
                 'status' => $status
@@ -32,6 +44,9 @@
         }
         public function delete_news($id)
         {
+            if (!$this->db->table_exists('feeds')) {
+                return false;
+            }
             $this->db->where('feedId', $id);
             $this->db->delete('feeds');
         }
